Move bcrypt tests
